fix for incorrect cast in string conversions

+++ tier1/utlbuffer.cpp
@@ -94,14 +94,14 @@ CUtlCStringConversion::CUtlCStringConversion( char nEscapeChar, const char *pDel
 	memset( m_pConversion, 0x0, sizeof(m_pConversion) );
 	for ( int i = 0; i < nCount; ++i )
 	{
-		m_pConversion[ static_cast<size_t>(pArray[i].m_pReplacementString[0]) ] = pArray[i].m_nActualChar;
+		m_pConversion[ static_cast<uint8>(pArray[i].m_pReplacementString[0]) ] = pArray[i].m_nActualChar;
 	}
 }
 
 // Finds a conversion for the passed-in string, returns length
 char CUtlCStringConversion::FindConversion( const char *pString, int *pLength )
 {
-	char c = m_pConversion[ static_cast<size_t>(pString[0]) ];
+	char c = m_pConversion[ static_cast<uint8>(pString[0]) ];
 	*pLength = (c != '\0') ? 1 : 0;
 	return c;
 }
@@ -124,7 +124,7 @@ CUtlCharConversion::CUtlCharConversion( char nEscapeChar, const char *pDelimiter
 	for ( int i = 0; i < nCount; ++i )
 	{
 		m_pList[i] = pArray[i].m_nActualChar;
-		ConversionInfo_t &info = m_pReplacements[ static_cast<size_t>(m_pList[i]) ];
+		ConversionInfo_t &info = m_pReplacements[ static_cast<uint8>(m_pList[i]) ];
 		Assert( info.m_pReplacementString == 0 );
 		info.m_pReplacementString = pArray[i].m_pReplacementString;
 		info.m_nLength = Q_strlen( info.m_pReplacementString );
@@ -160,12 +160,12 @@ int CUtlCharConversion::GetDelimiterLength() const
 //-----------------------------------------------------------------------------
 const char *CUtlCharConversion::GetConversionString( char c ) const
 {
-	return m_pReplacements[ static_cast<size_t>(c) ].m_pReplacementString;
+	return m_pReplacements[ static_cast<uint8>(c) ].m_pReplacementString;
 }
 
 int CUtlCharConversion::GetConversionLength( char c ) const
 {
-	return m_pReplacements[ static_cast<size_t>(c) ].m_nLength;
+	return m_pReplacements[ static_cast<uint8>(c) ].m_nLength;
 }
 
 int CUtlCharConversion::MaxConversionLength() const
@@ -181,9 +181,9 @@ char CUtlCharConversion::FindConversion( const char *pString, int *pLength )
 {
 	for ( int i = 0; i < m_nCount; ++i )
 	{
-		if ( !Q_strcmp( pString, m_pReplacements[ static_cast<size_t>(m_pList[i]) ].m_pReplacementString ) )
+		if ( !Q_strcmp( pString, m_pReplacements[ static_cast<uint8>(m_pList[i]) ].m_pReplacementString ) )
 		{
-			*pLength = m_pReplacements[ static_cast<size_t>(m_pList[i]) ].m_nLength;
+			*pLength = m_pReplacements[ static_cast<uint8>(m_pList[i]) ].m_nLength;
 			return m_pList[i];
 		}
 	}
